+++ src/main/java/gift/web/dto/response/ErrorResponse.java
@@ -1,7 +1,7 @@
 package gift.web.dto.response;
 
 import gift.web.validation.exception.CustomException;
-import gift.web.validation.exception.code.ErrorCode;
+import gift.web.validation.exception.code.ErrorStatus;
 import java.time.LocalDateTime;
 import org.springframework.validation.BindingResult;
 
@@ -27,8 +27,8 @@ public class ErrorResponse {
         return new ErrorResponse(-40010, "INVALID_PARAMETER", bindingResult.getFieldError().getDefaultMessage());
     }
 
-    public static ErrorResponse of(ErrorCode errorCode, String description) {
-        return new ErrorResponse(errorCode.getCode(), errorCode.getCategory().getDescription(), description);
+    public static ErrorResponse of(ErrorStatus errorStatus, String description) {
+        return new ErrorResponse(errorStatus.getCode(), errorStatus.getCategory().getDescription(), description);
     }
 
     public int getCode() {
