+++ src/main/java/gift/web/validation/exception/CustomException.java
@@ -1,6 +1,6 @@
 package gift.web.validation.exception;
 
-import gift.web.validation.exception.code.ErrorCode;
+import gift.web.validation.exception.code.ErrorStatus;
 
 /**
  * 커스텀 예외를 정의하고자 하는 경우 상속하여야 한다.
@@ -24,9 +24,9 @@ public abstract class CustomException extends RuntimeException {
     }
 
     /**
-     * 예외가 어떤 {@link ErrorCode} 를 가지는지 반환한다.
-     * @return {@link ErrorCode}
+     * 예외가 어떤 {@link ErrorStatus} 를 가지는지 반환한다.
+     * @return {@link ErrorStatus}
      */
-    public abstract ErrorCode getErrorCode();
+    public abstract ErrorStatus getErrorCode();
 
 }
