+++ src/main/java/gift/web/validation/exception/client/InvalidCredentialsException.java
@@ -1,8 +1,8 @@
 package gift.web.validation.exception.client;
 
-import static gift.web.validation.exception.code.ErrorCode.UNAUTHORIZED_INVALID_CREDENTIALS;
+import static gift.web.validation.exception.code.ErrorStatus.UNAUTHORIZED_INVALID_CREDENTIALS;
 
-import gift.web.validation.exception.code.ErrorCode;
+import gift.web.validation.exception.code.ErrorStatus;
 
 public class InvalidCredentialsException extends ClientException {
 
@@ -25,7 +25,7 @@ public class InvalidCredentialsException extends ClientException {
     }
 
     @Override
-    public ErrorCode getErrorCode() {
+    public ErrorStatus getErrorCode() {
         return UNAUTHORIZED_INVALID_CREDENTIALS;
     }
